==> 02_Pandas/13_function.ipynb <==
# 13. 함수적용
---
- pandas에서 데이터 변형을 수행할 때, apply()함수는 강력한 도구이다.
- 이함수를 사용하면 DataFrame 또는 Series의 각 요소에 대해 원하는 연산을 수행 할 수 있다.

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

#df['최근1년_방문빈도'] = df['최근1년_방문빈도'] + '회'

**데이터에 함수 적용 `apply()`**
- DataFrame이나 Series의 각 요소에 함수를 적용할 때 사용된다.
- 특히 열(Column) 단위나 행(row) 단위로 복잡한 연산을 적용할 때 유용하다.

def add_keyword(count):
    return str(count) + '회'

add_keyword(5)

df['최근1년_방문빈도'] = df['최근1년_방문빈도'].apply(add_keyword)

df

# 첫번째 문자가 대문자인지 아닌지 -> 대문자가 아니면 대문자로 변환
def capitalize(interest):
    if pd.notnull(interest):
        return interest.capitalize()
    return interest

df['주요관심사'] = df['주요관심사'].apply(capitalize)

df

**`merge()`**
- SQL의 JOIN과 유사하게 두개의 DataFrame을 특정 열을 기준으로 결합한다.

df1 = pd.DataFrame(
    {
        'CustomerID' : [1,2,3],
        'Product' : ['Laptop','Mounse','Keyboard']
    }
)

df1

df2 = pd.DataFrame(
    {
        'CustomerID' : [3,2,1],
        'Location' : ['USA','Canada','Korea']
    }
)
df2

df_merge = pd.merge(df1, df2, on='CustomerID')

df_merge

**`concat()`**
- DataFrame을 위아래또는 좌, 우로 결합할때 사용한다.
- 행을 추가할때는 `axis=0`, 열을추가할땐는 `axis=1`을 사용한다.

df1 = pd.DataFrame(
    {
        'A' : ['A0', 'A1'],
        'B' : ['B0','B1']
    }
)
df1

df2 = pd.DataFrame(
    {
        'A' : ['A2', 'A3'],
        'B' : ['B2','B3']
    }
)
df2

df_concat = pd.concat([df1, df2], axis=0)
df_concat

**`join()`**
- 두개의 DataFrame을 인덱스를 기준으로 결합 할 때 사용한다.

df1 = pd.DataFrame(
    {
        'A' : ['A0', 'A1'], 
    },
    index = ['K0','K1']
)
df1

df2 = pd.DataFrame(
    {
        'B' : ['B0','B1'],
    },
    index = ['K0','K1']
)
df2

df_join = df1.join(df2)

df_join
